Close pipes when the thread ends
The _PipeReaderThread() class creates two pipes to manage
input/output, and uses SELECT() to know when there is data to be
read. Unfortunately, those two pipes aren't closed when
the thread ends, which means that the FDs of the pipes of new
classes will grow up and up. The problem is that SELECT has a
limit in the FD value that can monitor, usually 1024, which
means that any FD with a value greater than that can't be
monitored by it, and will show the error if tried to:

    ValueError: filedescriptor out of range in select()

With small snaps this is not a problem because they don't create
so many objects of this kind as to reach the limit, but with
very big snaps (like Gnome-42-2204-SDK, with more than 50
binaries), it can be achieved.

This MR partially fixes this by closing the FDs when the thread
ends. Anyway, the best solution is to also use POLL instead of
SELECT, because it doesn't have that limit (this is implemented
in #104 , which should be merged along this one).

fix https://github.com/canonical/craft-cli/issues/102